Address PR notes - Address fixes - Clean import statements - Clean up file structure - Add Exception file - Remove unnecessary files 	* Did not make actual notes, but went through codebase and made my own adjustments/clean-up
