--- scripts/test.py
@@ -1,134 +1,0 @@
-from WellsFargoPdfDocument import WellsFargoPdfDocumentBaseClass
-
-import json
-
-
-def test():
-    PDF = WellsFargoPdfDocumentBaseClass(
-        "./scripts/test/document1.pdf", False, "checking"
-    )
-    received = json.loads(PDF.get_json())
-    expected = get_mocked_expected()
-    if received == expected:
-        print(f"\nSuccessful test ran!")
-
-
-def get_mocked_expected():
-    return [
-        {
-            "date": "7/27",
-            "description": "Mobile Deposit : Ref Number :511250576650",
-            "deposits": "200.00",
-        },
-        {
-            "date": "7/29",
-            "description": "Paypal Transfer 150729 5S4228Y3Ptpjn American Youth Leaders",
-            "deposits": "1,000.00",
-        },
-        {
-            "date": "7/30",
-            "description": "Online Transfer to American Youth Leadershi Business MarketRate Savings xxxxxx4326 Ref #Ibety26Qsc on 07/30/15",
-            "withdrawals": "1,200.00",
-        },
-        {
-            "date": "7/31",
-            "description": "Purchase authorized on 07/30 Wal-Mart Check Pri 866-925-2432TX S585210857879420 Card 5799",
-            "withdrawals": "14.32",
-        },
-        {
-            "date": "7/31",
-            "description": "Recurring Transfer to American Youth Leader Business MarketRate Savings Ref #Oper2Yy26P xxxxxx4326",
-            "withdrawals": "25.00",
-        },
-        {
-            "date": "8/3",
-            "description": "Online Transfer From American Youth Leaders Business MarketRate Savings xxxxxx4326 Ref #Iben5Yq228 on 08/03/15",
-            "deposits": "500.00",
-        },
-        {
-            "date": "8/4",
-            "description": "Purchase authorized on 08/03 Application User F 877-829-5500UT S005215653109050 Card 5799",
-            "withdrawals": "400.00",
-        },
-        {
-            "date": "8/5",
-            "description": "Mobile Deposit : Ref Number :311050073507",
-            "deposits": "100.00",
-        },
-        {
-            "date": "8/7",
-            "description": "Mobile Deposit : Ref Number :017070465425",
-            "deposits": "60.00",
-        },
-        {
-            "date": "8/11",
-            "description": "Mobile Deposit : Ref Number :609110908697",
-            "deposits": "180.00",
-        },
-        {
-            "date": "8/17",
-            "description": "Mobile Deposit : Ref Number :313150527194",
-            "deposits": "60.00",
-        },
-        {
-            "date": "8/17",
-            "description": "Mobile Deposit : Ref Number :513150527405",
-            "deposits": "60.00",
-        },
-        {
-            "date": "8/17",
-            "description": "Mobile Deposit : Ref Number :213150527080",
-            "deposits": "180.00",
-        },
-        {
-            "date": "8/17",
-            "description": "Mobile Deposit : Ref Number :709170701020",
-            "deposits": "90.00",
-        },
-        {
-            "date": "8/17",
-            "description": "Mobile Deposit : Ref Number :609170700825",
-            "deposits": "180.00",
-        },
-        {
-            "date": "8/17",
-            "description": "Mobile Deposit : Ref Number :509170700700",
-            "deposits": "270.00",
-        },
-        {
-            "date": "8/17",
-            "description": "Mobile Deposit : Ref Number :909170701246",
-            "deposits": "30.00",
-        },
-        {
-            "date": "8/18",
-            "description": "Mobile Deposit : Ref Number :109180870545",
-            "deposits": "270.00",
-        },
-        {
-            "date": "8/24",
-            "description": "Mobile Deposit : Ref Number :411220452573",
-            "deposits": "100.00",
-        },
-        {
-            "date": "8/24",
-            "description": "Mobile Deposit : Ref Number :211220452357",
-            "deposits": "180.00",
-        },
-        {
-            "date": "8/24",
-            "description": "Mobile Deposit : Ref Number :311220452470",
-            "deposits": "180.00",
-        },
-        {
-            "date": "8/25",
-            "description": "Mobile Deposit : Ref Number :410250802536",
-            "deposits": "180.00",
-        },
-    ]
-
-
-if "__main__" == __name__:
-    test()
-
-# @see {https://pdfminersix.readthedocs.io/en/latest/reference/composable.html#laparams}
